+++ lib/SILOptimizer/Utils/Local.cpp
@@ -1578,3 +1578,48 @@ void swift::hoistAddressProjections(Operand &Op, SILInstruction *InsertBefore,
   }
 }
 
+void StaticInitCloner::add(SILInstruction *InitVal) {
+  // Don't schedule an instruction twice for cloning.
+  if (NumOpsToClone.count(InitVal) != 0)
+    return;
+
+  ArrayRef<Operand> Ops = InitVal->getAllOperands();
+  NumOpsToClone[InitVal] = Ops.size();
+  if (Ops.empty()) {
+    // It's an instruction without operands, e.g. a literal. It's ready to be
+    // cloned first.
+    ReadyToClone.push_back(InitVal);
+  } else {
+    // Recursively add all operands.
+    for (const Operand &Op : Ops) {
+      add(cast<SingleValueInstruction>(Op.get()));
+    }
+  }
+}
+
+SingleValueInstruction *
+StaticInitCloner::clone(SingleValueInstruction *InitVal) {
+  assert(NumOpsToClone.count(InitVal) != 0 && "InitVal was not added");
+  // Find the right order to clone: all operands of an instruction must be
+  // cloned before the instruction itself.
+  while (!ReadyToClone.empty()) {
+    SILInstruction *I = ReadyToClone.pop_back_val();
+
+    // Clone the instruction into the SILGlobalVariable
+    visit(I);
+
+    // Check if users of I can now be cloned.
+    for (SILValue result : I->getResults()) {
+      for (Operand *Use : result->getUses()) {
+        SILInstruction *User = Use->getUser();
+        if (NumOpsToClone.count(User) != 0 && --NumOpsToClone[User] == 0)
+          ReadyToClone.push_back(User);
+      }
+    }
+  }
+  assert(ValueMap.count(InitVal) != 0 &&
+         "Could not schedule all instructions for cloning");
+  return cast<SingleValueInstruction>(ValueMap[InitVal]);
+}
+
+
