+++ test/SILOptimizer/static_arrays.swift
@@ -8,13 +8,6 @@
 
 // Check if the optimizer is able to convert array literals to statically initialized arrays.
 
-// CHECK-LABEL: sil_global private @{{.*}}main{{.*}} = {
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 100
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 101
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 102
-// CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}}, {{[^,]*}})
-// CHECK-NEXT:  }
-
 // CHECK-LABEL: outlined variable #0 of arrayLookup(_:)
 // CHECK-NEXT:  sil_global private @{{.*}}arrayLookup{{.*}} = {
 // CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 10
@@ -57,11 +50,10 @@
 // CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}})
 // CHECK-NEXT:  }
 
-// CHECK-LABEL: outlined variable #0 of overwriteLiteral(_:)
-// CHECK-NEXT:  sil_global private @{{.*}}overwriteLiteral{{.*}} = {
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 1
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 2
-// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 3
+// CHECK-LABEL: sil_global private @{{.*}}main{{.*}} = {
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 100
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 101
+// CHECK-DAG:     integer_literal $Builtin.Int{{[0-9]+}}, 102
 // CHECK:         object {{.*}} ({{[^,]*}}, [tail_elems] {{[^,]*}}, {{[^,]*}}, {{[^,]*}})
 // CHECK-NEXT:  }
 
@@ -120,18 +112,6 @@ public func storeArray() {
   gg = [227, 228]
 }
 
-// CHECK-LABEL: sil {{.*}}overwriteLiteral{{.*}} : $@convention(thin) (Int) -> @owned Array<Int> {
-// CHECK:   global_value @{{.*}}overwriteLiteral{{.*}}
-// CHECK:   is_unique
-// CHECK:   store
-// CHECK:   return
-@inline(never)
-func overwriteLiteral(_ x: Int) -> [Int] {
-  var a = [ 1, 2, 3 ]
-  a[x] = 0
-  return a
-}
-
 struct Empty { }
 
 // CHECK-LABEL: sil {{.*}}arrayWithEmptyElements{{.*}} : $@convention(thin) () -> @owned Array<Empty> {
@@ -155,10 +135,6 @@ print(gg!)
 storeArray()
 // CHECK-OUTPUT-NEXT: [227, 228]
 print(gg!)
-// CHECK-OUTPUT-NEXT: [0, 2, 3]
-print(overwriteLiteral(0))
-// CHECK-OUTPUT-NEXT: [1, 0, 3]
-print(overwriteLiteral(1))
